Remove the ToolSettingsService pass-through in favor of its port
Every ToolSettingsService method forwarded 1:1 to IToolSettingsPort
without adding any logic, so consumers now depend on the port directly
and the composition root constructs ToolSettingsRepository against it.
The interface moves to its own file and the Domain-residency guard test
now pins IToolSettingsPort instead of the deleted wrapper.
