+++ Packages/src/Editor/Application/UnityCliLoopToolRegistrar.cs
@@ -16,7 +16,7 @@ namespace io.github.hatayama.UnityCliLoop.Application
     {
         private readonly IInternalToolNameProvider _internalToolNameProvider;
         private readonly UnityCliLoopToolExecutionService _toolExecutionService;
-        private readonly ToolSettingsService _toolSettingsService;
+        private readonly IToolSettingsPort _toolSettingsService;
         private readonly Func<IReadOnlyList<IUnityCliLoopTool>> _toolDiscovery;
         private UnityCliLoopToolRegistry _sharedRegistry;
 
@@ -24,7 +24,7 @@ namespace io.github.hatayama.UnityCliLoop.Application
 
         internal UnityCliLoopToolRegistrarService(
             IInternalToolNameProvider internalToolNameProvider,
-            ToolSettingsService toolSettingsService,
+            IToolSettingsPort toolSettingsService,
             UnityCliLoopToolExecutionService toolExecutionService,
             Func<IReadOnlyList<IUnityCliLoopTool>> toolDiscovery)
         {
