+++ Packages/src/Editor/Domain/IToolSettingsPort.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: ea90ee82c38204d35b059830b6c8d8b3
+guid: 70f526f69d014dd7bb8e56404e62741c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+  userData:
+  assetBundleName:
+  assetBundleVariant:
