+++ Packages/src/Editor/Domain/ToolSettingsToolLinkPolicy.cs
@@ -24,7 +24,7 @@ namespace io.github.hatayama.UnityCliLoop.Domain
             return !PausePointAuxiliaryToolNames.Contains(toolName);
         }
 
-        internal static bool IsToolEnabled(string toolName, ToolSettingsService toolSettingsService)
+        internal static bool IsToolEnabled(string toolName, IToolSettingsPort toolSettingsService)
         {
             Debug.Assert(!string.IsNullOrEmpty(toolName), "toolName must not be null or empty");
             Debug.Assert(toolSettingsService != null, "toolSettingsService must not be null");
